Store errors inside working directory + fixes
This commit includes:
- Renaming the `knife_ec_error_handler` to `ec_error_handler`
- Replacing the Error handler file from a temporal file to a persistent
one inside `{work_dir}/errors`, depending whether we are backing up or
restoring is the name of the file we will generate with a timestamp.
- Include the content of `server_api_class_eval.rb` inside
`ec_error_handler` since we are just patching classes that will add
exceptions and rescues for error handling.
- A few comments with more explanation
- Multiple tiny fixes
