+++ .gitignore
@@ -2,3 +2,4 @@
 vendor/
 .bundle
 coverage
+*.gem
